refactor(discord): extract authorization cluster into DiscordAuthorizationMixin (god-file Phase 3)
DiscordAdapter is 7,092 lines and 179 methods in a 9,841-line module. This
lifts the authorization cluster out to plugins/platforms/discord/authz_mixin.py,
the same mechanical mixin extraction that produced gateway/authz_mixin.py.

Moved verbatim (475 lines):

  _discord_channel_ids_allowed    _reject_slash
  _is_pairing_approved_user       _notify_unauthorized_slash
  _is_allowed_user                _resolve_allowed_usernames
  _warn_if_fail_closed_default    _evaluate_slash_authorization
  _check_slash_authorization      _read_dm_role_auth_guild (module helper)

plus the section comment documenting why the slash surface carries its own
gate. _read_dm_role_auth_guild has exactly one caller, _is_allowed_user, and
travels with it.

Behavior-neutral. Every body is byte-identical to what it replaced; the only
edits are the class declaration, one import, and the new module header. The
mixin precedes BasePlatformAdapter in the bases, so resolution order is what
it was when these methods sat on the class. logger is bound by explicit name
rather than __name__, so records keep the name plugins.platforms.discord.adapter
and getLogger returns the adapter's own logger object; discord is imported
under the same ImportError guard and resolves to the same module object, so
tests patching attributes on it still take effect.

adapter.py 9,841 -> 9,340 lines.
